refactor: runtime.py 1476→517줄 — 5-module OpenClaw-style decomposition
core/runtime_wiring/ 패키지 신설, 4개 모듈로 wiring 로직 추출:
- bootstrap.py (345줄): hooks, memory, session, config_watcher, task, prompt, plugin registry
- infra.py (228줄): policies, tools, LLM adapters, auth, lanes
- automation.py (261줄): L4.5 9 components + hook wiring (table-driven logger hooks)
- adapters.py (243줄): MCP signal/notification/calendar/gateway

GeodeRuntime 클래스 + dataclass + instance methods만 runtime.py에 잔류.
기존 import 경로 backward compat 유지 (re-export aliases).
Pure refactor — 동작 변경 없음, 3232 tests pass.
